# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  tags: string[];
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  githubUrl: string | null;
  /** @nullable */
  liveUrl: string | null;
  viewCount: number;
  displayOrder: number;
  createdAt: string;
}

export interface SubmitContactBody {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  message: string;
  createdAt: string;
}

export interface PortfolioStats {
  totalProjects: number;
  totalViews: number;
  totalMessages: number;
}

export interface ErrorResponse {
  error: string;
}