# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all portfolio projects ordered by display order
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  liveUrl: zod.string().nullable(),
  viewCount: zod.number(),
  displayOrder: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  liveUrl: zod.string().nullable(),
  viewCount: zod.number(),
  displayOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Increment project view count
 */
export const IncrementProjectViewParams = zod.object({
  id: zod.coerce.number(),
});

export const IncrementProjectViewResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  liveUrl: zod.string().nullable(),
  viewCount: zod.number(),
  displayOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Submit a contact message
 */

export const SubmitContactBody = zod.object({
  name: zod.string().min(1),
  email: zod.string().email(),
  message: zod.string().min(1),
});

/**
 * Returns aggregate portfolio statistics
 * @summary Get portfolio statistics
 */
export const GetStatsResponse = zod.object({
  totalProjects: zod.number(),
  totalViews: zod.number(),
  totalMessages: zod.number(),
});