# lib/api-zod/src/generated/types/portfolioStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioStats {
  totalProjects: number;
  totalViews: number;
  totalMessages: number;
}