# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  title: string;
  description: string;
  tags: string[];
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  githubUrl: string | null;
  /** @nullable */
  liveUrl: string | null;
  viewCount: number;
  displayOrder: number;
  createdAt: Date;
}